+++ src/equation/roll.rs
@@ -4,18 +4,18 @@ use super::Token;
 use crate::equation;
 use rand::Rng;
 
-pub(super) fn process(equation: &Equation, ty: equation::RollType) -> i64 {
+pub(super) fn process(equation: &Equation, ty: equation::RollType) -> i32 {
     // todo!();
-    let mut stack: Vec<i64> = Vec::with_capacity(equation.compiled_equation.len());
+    let mut stack: Vec<i32> = Vec::with_capacity(equation.compiled_equation.len());
     for token in &equation.compiled_equation {
         match *token {
-            Token::Operand(value) => stack.push(value as i64),
+            Token::Operand(value) => stack.push(value as i32),
             Token::Dice(die) => match ty {
                 equation::RollType::Default => stack.push(roll(die)),
-                equation::RollType::Low => stack.push(die.number as i64),
-                equation::RollType::High => stack.push((die.number * die.sides) as i64),
+                equation::RollType::Low => stack.push(die.number as i32),
+                equation::RollType::High => stack.push((die.number * die.sides) as i32),
                 equation::RollType::Average => {
-                    stack.push((die.number as f32 * (die.sides as f32 / 2.0 + 0.5)) as i64)
+                    stack.push((die.number as f32 * (die.sides as f32 / 2.0 + 0.5)) as i32)
                 }
             },
             Token::Plus => {
@@ -49,7 +49,7 @@ pub(super) fn process(equation: &Equation, ty: equation::RollType) -> i64 {
                     0 => stack.push(1),
                     1 => stack.push(d),
                     _ => {
-                        let mut b: i64 = 1;
+                        let mut b: i32 = 1;
                         for _n in 0..ex {
                             b *= d;
                         }
@@ -62,13 +62,13 @@ pub(super) fn process(equation: &Equation, ty: equation::RollType) -> i64 {
     }
     stack.pop().unwrap()
 }
-fn roll(die: Die) -> i64 {
+fn roll(die: Die) -> i32 {
     // todo!();
     let mut rng = rand::thread_rng();
-    let mut current: i64 = 0;
+    let mut current: i32 = 0;
     //let die = dice.unwrap();
     for _n in 0..die.number {
-        current += rng.gen_range(1..die.sides + 1) as i64;
+        current += rng.gen_range(1..die.sides + 1) as i32;
     }
     current
 }
